requirements.txt
alembic==1.4.2
appdirs==1.4.3
astroid==2.3.3
blinker==1.4
blis==0.4.1
bokeh==2.2.0
Bootstrap-Flask==1.4
boto==2.49.0
boto3==1.11.7
botocore==1.14.7
Brotli==1.0.7
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
cymem==2.0.3
dash==1.14.0
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
distlib==0.3.0
docutils==0.15.2
en-core-web-lg==2.2.5
en-core-web-sm==2.2.5
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Compress==1.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
gensim==3.8.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.14.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
matplotlib==3.3.0
mccabe==0.6.1
murmurhash==1.0.2
numpy==1.18.1
packaging==20.4
pandas==0.25.3
Pillow==7.2.0
plac==1.1.3
plotly==4.9.0
preshed==3.0.2
pylint==2.4.4
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2019.3
pywin32==228
PyYAML==5.3.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.1
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
smart-open==1.9.0
spacy==2.2.3
SQLAlchemy==1.3.18
srsly==1.0.1
thinc==7.3.1
tornado==6.0.4
tqdm==4.41.1
typing-extensions==3.7.4.3
urllib3==1.25.7
virtualenv==20.0.18
wasabi==0.6.0
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.3.3
